Update installer for config.cson so it doesn’t install config.cson on each run of the installer

+++ includes/install/atom.sh
@@ -35,18 +35,28 @@ fi
 # Install Atom’s config.cson
 #####################################################################################
 if [[ -e "$HOME/.atom/$ATOM_CONFIG" ]]; then
-   inform  "Atom’s $ATOM_CONFIG file exists." true
-   pause_and_warn
+   if [[ -e "$HOME/.atom/backup$ATOM_CONFIG" ]]; then
+      inform  "Atom’s ~/.atom/backup$ATOM_CONFIG file exists. Continuing..." true
+   else
+      inform  "Atom’s $ATOM_CONFIG file exists." true
+      pause_and_warn
 
-   inform "Renaming config.cson to .atom/backup.config.cson..."
-   mv -v "$HOME/.atom/$ATOM_CONFIG" "$HOME/.atom/backup$ATOM_CONFIG"
+      inform "Renaming $ATOM_CONFIG to .atom/backup$ATOM_CONFIG..."
+      mv -v "$HOME/.atom/$ATOM_CONFIG" "$HOME/.atom/backup$ATOM_CONFIG"
+
+      inform "Downloading Atom’s $ATOM_CONFIG..."
+      curl -O https://raw.githubusercontent.com/code-warrior/web-dev-env-config-files/master/atom/config.cson
+
+      inform "Installing Atom’s $ATOM_CONFIG..."
+      mv -v config.cson "$HOME/.atom/"
+   fi
+else
+   inform "Downloading Atom’s $ATOM_CONFIG..."
+   curl -O https://raw.githubusercontent.com/code-warrior/web-dev-env-config-files/master/atom/config.cson
+
+   inform "Installing Atom’s $ATOM_CONFIG..."
+   mv -v config.cson "$HOME/.atom/"
 fi
-
-inform "Downloading Atom’s $ATOM_CONFIG..."
-curl -O https://raw.githubusercontent.com/code-warrior/web-dev-env-config-files/master/atom/config.cson
-
-inform "Installing Atom’s $ATOM_CONFIG..."
-mv -v config.cson "$HOME/.atom/"
 
 #####################################################################################
 # Install Atom packages
